removing extra files, fixing logging config by reorging project. Adding repr highlighter
