--- src/canhydro/import_option.py
@@ -1,9 +1,0 @@
-
-
-from importlib.util import find_spec
-
-def _try_import(package_name):
-    if find_spec(package_name):
-        return True
-    else:
-        return False
