+++ src/canhydro/user_def_config.toml
@@ -54,7 +54,7 @@ Source = "https://github.com/wischmcj/canopyHydrodynamics/"
 [directories]
 root_dir = "./"
 input_dir = "./data/input/"
-output_dir = "./data/output_dir/"
+output_dir = "./data/output/"
 test_input_dir = "./data/test/"
 
 [tool.ruff]
