+++ src/log_utils.py
@@ -1,0 +1,7 @@
+import logging
+import logging.config
+import yaml
+
+with open('logging_config.yml', 'rt') as f:
+    config = yaml.safe_load(f.read())
+    logging.config.dictConfig(config)
